Restructured project to match standards.
