--- GameBoard.h
@@ -1,445 +1,0 @@
-#pragma once
-#include <vector>
-#include <raylib.h>
-#include <random>
-#include <type_traits>
-
-struct IntVector2
-{
-	int x, y;
-
-	bool operator==(const IntVector2& other) const {
-		return x == other.x && y == other.y;
-
-	}
-};
-
-class GameBoard
-{
-public:
-	const enum AnchorPoints
-	{
-		TOP_LEFT,
-		TOP_MIDDLE,
-		TOP_RIGHT,
-
-		MIDDLE_LEFT,
-		MIDDLE,
-		MIDDLE_RIGHT,
-
-		BOTTOM_LEFT,
-		BOTTOM_MIDDLE,
-		BOTTOM_RIGHT,
-	};
-
-	class Drawable
-	{
-	private:
-		IntVector2 m_margin = { 0,0 };
-		IntVector2 m_dimensions = { 0,0 };
-		IntVector2 m_positionOnScreen = { 0,0 };
-		IntVector2 m_coords = { 0,0 };
-	
-	public:
-		Drawable(const IntVector2 dimensions, const IntVector2 margin)
-			: m_dimensions(dimensions), m_margin(margin)
-		{
-		}
-
-		virtual void Render() const = 0;
-
-		virtual int GetWidth() const
-		{
-			return m_dimensions.x;
-		}
-		virtual void SetWidth(const int width)
-		{
-			m_dimensions.x = width;
-		};
-
-		virtual int GetHeight() const
-		{
-			return m_dimensions.y;
-		}
-		virtual void SetHeight(const int height)
-		{
-			m_dimensions.y = height;
-		}
-
-		virtual int GetMarginWidth() const
-		{
-			return m_margin.x;
-		}
-		virtual void SetMarginWidth(const int marginWidth)
-		{
-			m_margin.x = marginWidth;
-		}
-
-		virtual int GetMarginHeight() const
-		{
-			return m_margin.y;
-		}
-		virtual void SetMarginHeight(const int marginHeight)
-		{
-			m_margin.y = marginHeight;
-		}
-
-		virtual void SetPositionOnScreen(const int x, const int y)
-		{
-			m_positionOnScreen = { x,y };
-		};
-		virtual IntVector2 GetPositionOnScreen() const
-		{
-			return m_positionOnScreen;
-		}
-	
-		virtual void SetGridCoords(const IntVector2 coords)
-		{
-			m_coords = coords;
-		}
-		virtual IntVector2 GetGridCoords() const
-		{
-			return m_coords;
-		}
-	};
-
-	class DrawableTexture : public Drawable
-	{
-	private:
-		Texture2D m_renderedTexture;
-
-	public:
-		DrawableTexture(const char* textureFilePath, const IntVector2 pixalDimensions, const IntVector2 margin)
-			:Drawable(pixalDimensions, margin),
-			m_renderedTexture(LoadTexture(textureFilePath))
-		{
-		}
-
-		Texture2D GetTexture() const
-		{
-			return m_renderedTexture;
-		}
-
-		void SetTexture(const char* textureFilePath)
-		{
-			if (!FileExists(textureFilePath)) throw std::invalid_argument("File does not exist.");
-
-			m_renderedTexture = LoadTexture(textureFilePath);
-		}
-
-		void Render() const override
-		{
-			IntVector2 positionOnScreen = GetPositionOnScreen();
-			float scaleY = GetHeight() / m_renderedTexture.height;
-			float scaleX = GetWidth() / m_renderedTexture.width;
-
-			float scale = std::min(scaleX, scaleY);
-
-			DrawTextureEx(m_renderedTexture, { (float)positionOnScreen.x, (float)positionOnScreen.y }, 0, scale, WHITE);
-		};
-	};
-
-	class Text : public Drawable
-	{
-	private:
-		std::string m_text;
-		int m_fontSize;
-		Font m_font = GetFontDefault();
-		Color m_colour;
-		IntVector2 m_positionOnScreen = { 0,0 };
-		AnchorPoints m_anchorPoint = AnchorPoints::TOP_LEFT;
-		
-	public:
-
-		Text(std::string text, int fontSize, Color colour)
-			:Drawable(IntVector2{ MeasureText(text.c_str(), fontSize), 10 }, IntVector2{0,0}),
-			m_text(text), m_fontSize(fontSize), m_colour(colour)
-		{
-		}
-
-		std::string GetText() const
-		{
-			return m_text;
-		}
-		void SetText(const std::string text)
-		{
-			m_text = text;
-		}
-
-		int GetFontSize() const
-		{
-			return m_fontSize;
-		}
-		void SetFontSize(const int fontSize)
-		{
-			m_fontSize = fontSize;
-		}
-
-		int GetWidth() const override
-		{
-			return MeasureTextEx(m_font, m_text.c_str(), m_fontSize, 0).x;
-		}
-
-		int GetHeight() const override
-		{
-			return MeasureTextEx(m_font, m_text.c_str(), m_fontSize, 0).y;
-		}
-
-		AnchorPoints GetAnchorPoint() const
-		{
-			return m_anchorPoint;
-		}
-		void SetAnchorPoint(const AnchorPoints anchorpoint)
-		{
-			m_anchorPoint = anchorpoint;
-			SetPositionOnScreen(m_positionOnScreen.x, m_positionOnScreen.y);
-		}
-
-		void SetPositionOnScreen(int x, int y) override
-		{
-			const int height = GetHeight();
-			const int width = GetWidth();
-			IntVector2 offset = { 0,0 };
-
-			switch (m_anchorPoint)
-			{
-			case GameBoard::TOP_MIDDLE:
-				offset.x = width / 2;
-				break;
-
-			case GameBoard::TOP_RIGHT:
-				offset.x = width;
-				break;
-
-			case GameBoard::MIDDLE_LEFT:
-				offset.y = height / 2;
-				break;
-
-			case GameBoard::MIDDLE:
-				offset.x = width / 2;
-				offset.y = height / 2;
-				break;
-
-			case GameBoard::MIDDLE_RIGHT:
-				offset.x = width;
-				offset.y = height / 2;
-				break;
-
-			case GameBoard::BOTTOM_LEFT:
-				offset.y = height;
-				break;
-
-			case GameBoard::BOTTOM_MIDDLE:
-				offset.x = width / 2;
-				offset.y = height;
-				break;
-
-			case GameBoard::BOTTOM_RIGHT:
-				offset.x = width;
-				offset.y = height;
-				break;
-
-			default:
-				break;
-			}
-		
-			x -= offset.x;
-			y -= offset.y;
-
-			m_positionOnScreen = { x,y };
-		}
-
-		Color GetColour() const
-		{
-			return m_colour;
-		}
-		void SetColour(const Color colour)
-		{
-			m_colour = colour;
-		}
-	
-		void Render() const override
-		{
-			DrawText(m_text.c_str(), m_positionOnScreen.x, m_positionOnScreen.y, m_fontSize, m_colour);
-		}
-	};
-
-	template <typename T_Entity = Drawable >
-	class Grid
-	{
-		
-	protected:
-
-		AnchorPoints m_anchorPoint = AnchorPoints::TOP_LEFT;
-		typedef std::vector<std::vector<T_Entity>> T_Grid;
-		T_Grid m_grid;
-
-	protected:
-		int GenerateRandomInteger(const int min, const int max) {
-			// Create a random device and a Mersenne Twister random number generator
-			std::random_device rd;
-			std::mt19937 gen(rd());
-
-			// Define a distribution in the specified range [min, max]
-			std::uniform_int_distribution<int> dis(min, max);
-
-			// Generate and return the random number
-			return dis(gen);
-		}
-
-		T_Grid GenerateBoard(const IntVector2 dimensions, T_Entity sampleSquare)
-		{
-			static_assert(std::is_base_of<Drawable, T_Entity>::value, "T_Entity must derive from Gameboard::Drawable");
-
-			if (dimensions.x <= 0 || dimensions.y <= 0) {
-				throw std::invalid_argument("Board dimensions must be positive");
-			}
-
-			T_Grid board;
-
-			// Populate board with tiles.
-			for (int y = 0; y < dimensions.y; y++)
-			{
-				std::vector<T_Entity> row;
-
-				for (int x = 0; x < dimensions.x; x++)
-				{
-					sampleSquare.SetGridCoords(IntVector2{ x, y });
-					row.push_back(sampleSquare);
-				}
-
-				board.push_back(row);
-			}
-
-			return board;
-		}
-
-		IntVector2 GetBoardPixelDimensions() const
-		{
-			if (m_grid.empty() || m_grid[0].empty()) {
-				throw std::runtime_error("Empty grid");
-			}
-
-			IntVector2 dimensions;
-			T_Entity square = m_grid[0][0];
-
-			dimensions.x = m_grid[0].size() * square.GetWidth() + (m_grid[0].size() - 1) * square.GetMarginWidth();
-			dimensions.y = m_grid.size() * square.GetHeight() + (m_grid.size() - 1) * square.GetMarginHeight();
-
-			return dimensions;
-		}
-
-		virtual bool ShouldRenderEntity(const IntVector2 coords) const
-		{
-			return true;
-		}
-
-	public:
-		Grid(const IntVector2 dimensions, const T_Entity sampleEntity, const AnchorPoints anchorPoint, const IntVector2 position)
-		{
-			m_grid = GenerateBoard(dimensions, sampleEntity);
-			SetAnchorPoint(anchorPoint);
-			SetPositionsOnScreen(position);
-		}
-
-		void SetAnchorPoint(const AnchorPoints anchorPoint) {
-			m_anchorPoint = anchorPoint;
-		}
-
-		void SetPositionsOnScreen(const IntVector2 position)
-			// Function adds position of each square on the screen to it's attributes so other methods can access it.
-		{
-			IntVector2 offset = { 0,0 }; // Default for top left anchor point;
-			IntVector2 pixelDimensions = GetBoardPixelDimensions();
-
-			switch (m_anchorPoint)
-			{
-			case TOP_MIDDLE: offset.x = int(pixelDimensions.x / 2); break;
-			case TOP_RIGHT: offset.x = pixelDimensions.x; break;
-			case MIDDLE_LEFT: offset.y = int(pixelDimensions.y / 2); break;
-
-			case MIDDLE:
-				offset.x = int(pixelDimensions.x / 2);
-				offset.y = int(pixelDimensions.y / 2);
-				break;
-
-			case MIDDLE_RIGHT:
-				offset.x = pixelDimensions.x;
-				offset.y = int(pixelDimensions.y / 2);
-				break;
-
-			case BOTTOM_LEFT:
-				offset.y = pixelDimensions.y; 
-				break;
-
-			case BOTTOM_MIDDLE:
-				offset.x = int(pixelDimensions.x / 2);
-				offset.y = pixelDimensions.y;
-				break;
-
-			case BOTTOM_RIGHT:
-				offset.x = pixelDimensions.x;
-				offset.y = pixelDimensions.y;
-				break;
-
-			default:
-				break;
-			}
-
-			for (int y = 0; y < m_grid.size(); y++)
-			{
-				for (int x = 0; x < m_grid[y].size(); x++)
-				{
-					T_Entity entity = m_grid[y][x];
-					int xPos = (x * (entity.GetMarginWidth() + entity.GetWidth())) + position.x - offset.x;
-					int yPos = (y * (entity.GetMarginHeight() + entity.GetHeight())) + position.y - offset.y;
-
-					m_grid[y][x].SetPositionOnScreen(xPos, yPos);
-				}
-			}
-		}
-
-		void DisplayGrid() const
-		{
-
-			for (int y = 0; y < m_grid.size(); y++)
-			{
-				for (int x = 0; x < m_grid[y].size(); x++)
-				{
-					if (ShouldRenderEntity(IntVector2{ x, y }))
-					{
-						m_grid[y][x].Render();
-					}
-				}
-			}
-		}
-
-		virtual void ProcessMouseInput()
-		{
-
-		}
-	
-		std::vector<T_Entity> GetNeighbours(const T_Entity& tile) const
-		{
-			std::vector<T_Entity> neighbours;
-			int x = tile.GetGridCoords().x;
-			int y =	tile.GetGridCoords().y;
-
-			const int dx[] = { -1, -1, -1,  0, 0,  1, 1, 1 };
-			const int dy[] = { -1,  0,  1, -1, 1, -1, 0, 1 };
-
-			for (int i = 0; i < 8; ++i) 
-			{
-				int newX = x + dx[i];
-				int newY = y + dy[i];
-				if (newX >= 0 && newX < m_grid[0].size() &&
-					newY >= 0 && newY < m_grid.size()) 
-				{
-					neighbours.push_back(m_grid[newY][newX]);
-				}
-			}
-
-			return neighbours;
-		}
-	};
-};
